+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - Migrate from tox to pipenv
   - Use tox for local testing only
   - Integrate pytest and setup.py
+  - Replace `release.sh` with `python setup.py upload`
 
 ### Removed